pyproject.toml
[project]
name = "codex-call-audit"
version = "0.1.0"
description = "Offline call quality assessment pipeline"
authors = [{ name = "Codex AI" }]
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.3.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
pipeline = [
    "faster-whisper>=1.0.0",
    "pyannote.audio>=3.0.0",
    "soundfile>=0.12.1",
    "torchaudio>=2.0.0",
    "librosa>=0.10.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
